checker/requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
flatten-dict==0.4.2
idna==3.4
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
soupsieve==2.4
urllib3==1.26.14
user-agent==0.1.10
selenium
selenium-wire
lxml
